N21-1053 move externalToolDO
