+++ apps/server/src/modules/tool/external-tool/domainobject/config/basic-tool-config.do.ts
