+++ apps/server/src/modules/tool/external-tool/domainobject/config/external-tool-config.do.ts
