+++ apps/server/src/modules/tool/external-tool/domainobject/config/index.ts
