+++ apps/server/src/modules/tool/external-tool/domainobject/config/oauth2-tool-config.do.ts
