+++ apps/server/src/modules/tool/external-tool/domainobject/custom-parameter.do.ts
@@ -1,4 +1,4 @@
-import { CustomParameterLocation, CustomParameterScope, CustomParameterType } from '@shared/domain';
+import { CustomParameterLocation, CustomParameterScope, CustomParameterType } from '../entity';
 
 export class CustomParameterDO {
 	name: string;
