+++ apps/server/src/modules/tool/external-tool/domainobject/external-tool.do.spec.ts
