+++ apps/server/src/modules/tool/external-tool/domainobject/index.ts
@@ -1,0 +1,3 @@
+export * from './external-tool.do';
+export * from './config';
+export * from './custom-parameter.do';
