+++ apps/server/src/modules/tool/external-tool/domainobject/types/index.ts
