+++ apps/server/src/modules/tool/external-tool/entity/config/basic-tool-config.ts
