+++ apps/server/src/modules/tool/external-tool/entity/config/external-tool-config.ts
