+++ apps/server/src/modules/tool/external-tool/entity/config/oauth2-tool-config.ts
