+++ apps/server/src/modules/tool/external-tool/entity/custom-parameter/custom-parameter-scope.enum.ts
