+++ apps/server/src/modules/tool/external-tool/entity/custom-parameter/custom-parameter-type.enum.ts
