+++ apps/server/src/modules/tool/external-tool/entity/external-tool.entity.spec.ts
