+++ apps/server/src/modules/tool/external-tool/entity/external-tool.entity.ts
@@ -1,6 +1,6 @@
 import { Embedded, Entity, Property, Unique } from '@mikro-orm/core';
-import { BasicToolConfig, Lti11ToolConfig, Oauth2ToolConfig } from '@shared/domain/entity/tools/external-tool/config';
-import { BaseEntityWithTimestamps } from '../../base.entity';
+import { BaseEntityWithTimestamps } from '@shared/domain';
+import { BasicToolConfig, Lti11ToolConfig, Oauth2ToolConfig } from './config';
 import { CustomParameter } from './custom-parameter';
 
 export type IExternalToolProperties = Readonly<Omit<ExternalTool, keyof BaseEntityWithTimestamps>>;
