+++ apps/server/src/modules/tool/external-tool/entity/index.ts
