+++ apps/server/src/modules/tool/external-tool/service/external-tool-parameter-validation.service.spec.ts
@@ -1,14 +1,14 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { Test, TestingModule } from '@nestjs/testing';
 import { ValidationError } from '@shared/common';
-import { CustomParameterScope, CustomParameterType } from '@shared/domain';
-import { CustomParameterDO, ExternalToolDO } from '@shared/domain/domainobject/tool';
 import {
 	customParameterDOFactory,
 	externalToolDOFactory,
 } from '@shared/testing/factory/domainobject/tool/external-tool.factory';
 import { ExternalToolService } from './index';
 import { ExternalToolParameterValidationService } from './external-tool-parameter-validation.service';
+import { CustomParameterScope, CustomParameterType } from '../entity';
+import { ExternalToolDO, CustomParameterDO } from '../domainobject';
 
 describe('ExternalToolParameterValidationService', () => {
 	let module: TestingModule;
