+++ apps/server/src/modules/tool/external-tool/uc/external-tool-configuration.uc.spec.ts
@@ -3,15 +3,7 @@ import { ObjectId } from '@mikro-orm/mongodb';
 import { ForbiddenException, UnauthorizedException } from '@nestjs/common';
 import { NotFoundException } from '@nestjs/common/exceptions/not-found.exception';
 import { Test, TestingModule } from '@nestjs/testing';
-import {
-	ContextExternalToolDO,
-	CustomParameterScope,
-	ExternalToolDO,
-	Page,
-	Permission,
-	SchoolExternalToolDO,
-	User,
-} from '@shared/domain';
+import { ContextExternalToolDO, Page, Permission, SchoolExternalToolDO, User } from '@shared/domain';
 import { contextExternalToolDOFactory, setupEntities, userFactory } from '@shared/testing';
 import { externalToolDOFactory, schoolExternalToolDOFactory } from '@shared/testing/factory/domainobject/tool';
 import { ICurrentUser } from '@src/modules/authentication';
@@ -19,8 +11,10 @@ import { Action, AuthorizableReferenceType, AuthorizationService } from '../../.
 import { ExternalToolConfigurationUc } from './external-tool-configuration.uc';
 import { ToolContextType } from '../../common/interface';
 import { ExternalToolService } from '../service';
-import { SchoolExternalToolService } from '../../school-external-tool/service/school-external-tool.service';
+import { SchoolExternalToolService } from '../../school-external-tool/service';
 import { ContextExternalToolService } from '../../context-external-tool/service';
+import { ExternalToolDO } from '../domainobject';
+import { CustomParameterScope } from '../entity';
 
 describe('ExternalToolConfigurationUc', () => {
 	let module: TestingModule;
