+++ apps/server/src/modules/tool/external-tool/uc/external-tool-configuration.uc.ts
@@ -1,14 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { NotFoundException } from '@nestjs/common/exceptions/not-found.exception';
-import {
-	ContextExternalToolDO,
-	CustomParameterScope,
-	EntityId,
-	Permission,
-	SchoolExternalToolDO,
-} from '@shared/domain';
+import { ContextExternalToolDO, EntityId, Permission, SchoolExternalToolDO } from '@shared/domain';
 import { Page } from '@shared/domain/domainobject/page';
-import { ExternalToolDO } from '@shared/domain/domainobject/tool';
 import { Action, AuthorizableReferenceType, AuthorizationService } from '@src/modules/authorization';
 import { ToolContextType } from '../../common/interface';
 import { ExternalToolService } from '../service';
@@ -16,6 +9,8 @@ import { SchoolExternalToolService } from '../../school-external-tool/service';
 import { ContextExternalToolService } from '../../context-external-tool/service';
 import { AvailableToolsForContext } from './dto';
 import { ContextTypeMapper } from '../../common/mapper';
+import { ExternalToolDO } from '../domainobject';
+import { CustomParameterScope } from '../entity';
 
 @Injectable()
 export class ExternalToolConfigurationUc {
