+++ apps/server/src/modules/tool/external-tool/uc/external-tool.uc.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { EntityId, ExternalToolConfigDO, ExternalToolDO, IFindOptions, Page, Permission, User } from '@shared/domain';
+import { EntityId, IFindOptions, Page, Permission, User } from '@shared/domain';
 import { AuthorizationService } from '@src/modules/authorization';
 import { ExternalToolSearchQuery } from '../../common/interface';
+import { ExternalToolConfigDO, ExternalToolDO } from '../domainobject';
 import { ExternalToolService, ExternalToolValidationService } from '../service';
 import { ExternalToolCreate, ExternalToolUpdate } from './dto';
