+++ apps/server/src/shared/domain/domainobject/tool/index.ts
@@ -1,11 +1,7 @@
-export * from './external-tool.do';
-export * from './custom-parameter.do';
 export * from './context-external-tool.do';
 export * from './custom-parameter-entry.do';
 export * from './school-external-tool.do';
 export * from './tool-configuration-status';
 export * from './school-external-tool-ref.do';
-export * from './config';
 export * from './context-ref';
 export * from './tool-reference';
-export * from './types';
